refactor(api/src/models/Category) changed filename and paths
