+++ api/src/app.js
@@ -2,8 +2,8 @@ const express = require('express');
 const cookieParser = require('cookie-parser');
 const morgan = require('morgan');
 const cors = require('cors');
-
 const setHeaders =  require('./setHeaders');
+const routes = require('./routes/index');
 
 const server = express();
